Indirect precodes using special indirection cell.
New indirection cell created for each method which is called from the same module.
MethodDesc has a slot which points to the indirection cell so that it's easy to find (relative pointer, optional slot after the main method desc and other optional slots).
FixupPrecode patching is modified to fix the indirection cell and leave the precode as is.

Doesn't work:
- PreStub and DoBackpatch use the stable entry point as the address to jump to after the prestub is done. Since we don't change that, it still ends up in the unchanged precode and goes through prestub again... endless loop
- There's a "try" at hacking around this by forcing the return of the pCode, this seems to break the endless loop, but crashes somewhere else - didn't debug in detail.