Allows patching of slots instead of precodes for native images.
Modifies method table and vtable chunks to be writeable.
Methods with precode are saved as methods with "no-repcode" and "no-stable entrypoint". This emulates temporary entry points.
Modified places in MD where we assume that no stable entry point means temporary entry point, which is no longer the case - for zapped ethods this is a valid state and the "temporary entry point" is the value in the MT slot.
Mark all precodes as RX only.

With this the system works while zapped precodes are never modified.