Two fixes to handle special methods.
Unboxing stub and com interop. The same problem in both cases, the code expects to have a stable entry point with a precode in these cases, but with the RX changes that's not the case.
So solving this by calling GetOrCreatePrecode to force runtime allocated precode.
Needs more work around patching all the possible slots with this.