Simplify the code change by not allowing SetTarget on zapped precodeds.
Turns out that zapped precodes can't have the SetTarget called on them with one exception the backpatch. In that case doing it is moot anyway as it would not do anything (we're setting it to the stable entry point which is already in the main method slot at that time).
So I removed the ability to call SetTarget on zapped precodes completely.