Repurpose ZapMethodEntryPoint as ZapMethodSlot
Instead of introducing new ZapMethodSlot and leaving the ZapMethodEntryPoint unused - rename ZapMethodEntryPoint to ZapMethodSlot - also the cache is reused. This is the actually purpose of this change since the cache uses both method handle and access flags for lookup (as oppose to the wrappers which only use the method handle).
This will be needed once we start replacing call indirections with direct calls for methods which can be called directly (Where the decision relies on the access flags as well).
Some minor cleanup.