Add the abitlity to store temporary entry points in NGEN on MD chunk.
Basically saves teh MD chunk with one additional pointer in front of it. That's where we will store the temporary entry points pointer (just like the runtime allocated MD chunks do).
In order to do this cleanly we need to be able to map the MD chunk pointer to an ZapInnerPtr node (and not the structure node which backs up the memory since that is starting eith the temporary entry point pointer).
This introduces the ZapInnerPtr nodes to lot of the DataImage Fixup and BindPointer infra. So adapt that infra to be able to handle inner ptr nodes correctly.