Save temporary entry points for all methods in a chunk
Change chunk method ordering to consider the need for precodes - with this we try to put methods which need precodes together and split chunks on that. So we should have chunks where all (or most) methods need precodes and chunk where none do.
Save precodes only for chunks where at least one method needs them.

Register all precodes as surrogates of the "new" MD, so that we can run fixup on them (easy lookup). Leave the current registration of precodes as surrogates of the "old" MD only for methods which need precodes. This is used to figure out if the method should use precode in its MT slot.