Remove unnecessary special casing for zapped methods
Remove the special code to deal with zapped precodes in GetOrCreatePrecode and SetStableEntryPoint. In both cases we can now rely on the GetTemporaryEntryPoint to work correctly. The only exceptrion is GetOrCreatePrecode which still has to enfore new prcode creation over zapped precode, regardless of the type of the zapped precode.
Also removed the special code in DoBackpatch as it can also rely on GetTemporaryEntryPoint.